Move wizard helper test fixtures in to base conftest and use for user assignment wizard tests
