+++ unit_tests/conftest.py
@@ -4,7 +4,12 @@ import pytest
 import os
 import copy
 
+from urllib.parse import urljoin, urlparse
+
 from django.conf import settings
+from django.urls import resolve
+
+from formtools.wizard.views import normalize_name
 
 from core import client
 from core.constants import OrganisationDocumentType
@@ -2350,3 +2355,60 @@ def mock_control_list_entries_get(requests_mock):
             "control_list_entries": [{"rating": "ML1a", "text": "some text"}, {"rating": "ML22b", "text": "some text"}]
         },
     )
+
+
+class WizardStepPoster:
+    def __init__(self, authorized_client, url):
+        self.authorized_client = authorized_client
+        self.url = url
+
+    def post(self, post_data, **kwargs):
+        return self.authorized_client.post(
+            self.url,
+            data=post_data,
+            **kwargs,
+        )
+
+
+class WizardStepGoto(WizardStepPoster):
+    def __call__(self, step_name):
+        return self.post(
+            {
+                "wizard_goto_step": step_name,
+            }
+        )
+
+
+@pytest.fixture
+def goto_step_factory(authorized_client):
+    def goto_step(url):
+        step_gotoer = WizardStepGoto(authorized_client, url)
+        return step_gotoer
+
+    return goto_step
+
+
+class WizardStepPost(WizardStepPoster):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        url = urljoin(self.url, urlparse(self.url).path)
+        match = resolve(url)
+        self.view_name = normalize_name(match.func.__name__)
+
+    def __call__(self, step_name, data, **kwargs):
+        return self.post(
+            {
+                f"{self.view_name}-current_step": step_name,
+                **{f"{step_name}-{key}": value for key, value in data.items()},
+            },
+            **kwargs,
+        )
+
+
+@pytest.fixture
+def post_to_step_factory(authorized_client):
+    def post_to_step(url):
+        step_poster = WizardStepPost(authorized_client, url)
+        return step_poster
+
+    return post_to_step
